refacto services + tests
