+++ config/locales/models/applicant.fr.yml
@@ -23,3 +23,4 @@ fr:
         role: Rôle
         address: Adresse
         email: Email
+        title: Civilité
