+++ config/locales/simple_form.fr.yml
@@ -1,3 +1,4 @@
+
 fr:
   simple_form:
     "yes": 'Oui'
